+++ Platforms/Android/MainApplication.cs
@@ -1,17 +1,15 @@
 ﻿using Android.App;
 using Android.Runtime;
-using EnlightenMAUI.UI;
 
-namespace EnlightenMAUI
+namespace EnlightenMAUI;
+
+[Application]
+public class MainApplication : MauiApplication
 {
-    [Application]
-    public class MainApplication : MauiApplication
+    public MainApplication(IntPtr handle, JniHandleOwnership ownership)
+        : base(handle, ownership)
     {
-        public MainApplication(IntPtr handle, JniHandleOwnership ownership)
-            : base(handle, ownership)
-        {
-        }
-
-        protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
     }
+
+    protected override MauiApp CreateMauiApp() => MauiProgram.CreateMauiApp();
 }