+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+- 2024-04-15 0.0.4
+    - added enlighten_logo
+    - AboutViewModel -> code-behind
+    - Data -> Clipboard
+    - Device -> {Bluetooth, Hardware}
+    - added Clipboard
+    - all tabs viewable
 - 2024-04-14 0.0.3
     - Telerik chart loading
     - imported from EnlightenMobile